Builds container image with code bundled in repo
Building a container image with potentially different code than the one included in this repository can cause a lot of confusion: for example, imagine forking the project, then running a local `docker build` / `podman build` only to discover that the code in your container image doesn't contain the changes you applied locally...

Instead, this commit ties the `Dockerfile` with the code actually included in the repo: not some centralized copy of the source code, but the actual local working copy.

The change from `cargo install radicle-cli` => `cargo install --path cli` comes from the instructions in the `README`.